+++ src/merge_candidates.cpp
@@ -7,7 +7,7 @@
 #include <cassert>
 #include <tuple>
 #include <filesystem>
-#include <parse_mums.hpp>
+#include <mumsio.hpp>
 
 using namespace std;
